tutorials: cleanup of tutorial 08, rename to 08CustomPythonSearcher
