+++ Tutorials/08CustomPythonSearcher/Artifacts/stdout.log
