+++ Tutorials/08CustomPythonSearcher/base.py
